requirements.txt
PySide6>=6.5.0
pytest>=7.0.0
python-dotenv>=1.0.0
watchdog>=3.0.0
psutil>=5.9.0